my-app/chatbot_api/requirements.txt
fastapi
uvicorn
requests
llama-index-core
llama-index-embeddings-huggingface
urllib3 